refactor(validator): fix package
