+++ feature/validator/src/commonMain/kotlin/com/neoutils/neoregex/feature/validator/state/TestCase.kt
@@ -1,7 +1,7 @@
 /*
  * NeoRegex.
  *
- * Copyright (C) 2025 <AUTHOR>.
+ * Copyright (C) 2025 Irineu A. Silva.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neoutils.neorefex.feature.validator.state
+package com.neoutils.neoregex.feature.validator.state
 
 import androidx.compose.animation.*
 import androidx.compose.animation.core.RepeatMode
